Fix incorrect deletion of events.
When an event is fired by a timeout during the execution a new event may be
added to the event list. If this new event has a timeout that lies in the past,
the event is added to the front of the list. So after the execution it must not
be the _first_ element that is deleted, but the _former first_ element.